requirements.txt
requests==2.31.0
cloudscraper==1.2.70
beautifulsoup4==4.12.2
python-dateutil==2.8.2
pymongo==4.6.1
urllib3==1.26.16
pytz==2023.3
python-dotenv==1.0.0
pysocks==1.7.1